requirements.txt
Django==3.2.9
requests==2.25.1
python-dotenv==0.19.1